Registry login and logout methods moved out from utils to separate modules.
