+++ aea/cli/login.py
@@ -22,7 +22,7 @@
 import click
 
 from aea.cli.registry.settings import AUTH_TOKEN_KEY
-from aea.cli.registry.utils import registry_login
+from aea.cli.registry.login import registry_login
 from aea.cli.utils.config import update_cli_config
