+++ aea/cli/registry/logout.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # ------------------------------------------------------------------------------
 #
-#   Copyright 2018-2019 Fetch.AI Limited
+#   Copyright 2018-2020 Fetch.AI Limited
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -16,29 +16,15 @@
 #   limitations under the License.
 #
 # ------------------------------------------------------------------------------
+"""Registry utils used for CLI login command."""
 
-"""Implementation of the 'aea logout' subcommand."""
-
-import click
-
-from aea.cli.registry.settings import AUTH_TOKEN_KEY
-from aea.cli.registry.logout import registry_logout
-from aea.cli.utils.config import update_cli_config
+from aea.cli.registry.utils import request_api
 
 
-@click.command(name="logout", help="Logout from Registry account.")
-def logout():
-    """Logout from Registry account command."""
-    click.echo("Logging out...")
-    do_logout()
-    click.echo("Successfully logged out.")
-
-
-def do_logout() -> None:
+def registry_logout() -> None:
     """
     Logout from Registry account.
 
-    :return: None.
+    :return: None
     """
-    registry_logout()
-    update_cli_config({AUTH_TOKEN_KEY: None})
+    request_api("POST", "/rest-auth/logout/")
